lib/features/home/presentation/screens/image_search/furniture_image_search/requirements.txt
torch==2.0.0
transformers==4.30.0
faiss-cpu==1.7.4
flask==2.3.0
flask-cors==4.0.0
mysql-connector-python==8.0.33
pillow==9.5.0
numpy==1.24.0